Refactor Retriever & Document Retriever Base Classes
# Refactor Retriever & Document Retriever Base Classes

Pull out all the document-related stuff from the base retriever. We'll want to use retriever implementations for structured data retrieved from other data stores (e.g. CSV for demo, postgres and other dbs)
